Move keyword counting out of the ASTVisitor and into a Lexer
