+++ src/MetricUnit.cpp
@@ -267,7 +267,7 @@ const std::set<MetricType_e> m_keywordMetrics = {
 	METRIC_TYPE_VARIABLE_FN_AUTO,
 	METRIC_TYPE_BREAK,
 	METRIC_TYPE_CASE,
-//	METRIC_TYPE_CHAR,
+	METRIC_TYPE_CHAR,
 	METRIC_TYPE_CONST,
 	METRIC_TYPE_CONTINUE,
 	METRIC_TYPE_DEFAULT,
