+++ src/MetricUnit.hpp
@@ -53,7 +53,8 @@ typedef enum
 
 typedef enum {
 	METRIC_UNIT_PROCESS_AST,
-	METRIC_UNIT_PROCESS_LEX,
+	METRIC_UNIT_PROCESS_LEX_UNEXPANDED,
+	METRIC_UNIT_PROCESS_LEX_EXPANDED,
 	METRIC_UNIT_PROCESS_MAX
 } MetricUnitProcessingType_e;
